requirements.txt
alembic==1.14.0
bcrypt==4.2.1
certifi==2024.12.14
greenlet==3.1.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
psycopg2-binary==2.9.10
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.9.4
sentry-sdk==2.19.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
